requirements.txt
playwright==1.49.0
sentence-transformers==3.3.1
scikit-learn==1.5.2
pandas==2.2.3
supabase==2.10.0
python-dotenv==1.0.1
anthropic==0.40.0
pytest==9.0.2
